Remove set-target.sh
set-target.sh is an script that attempts to figure out the right go.mod
dependencies, be it for upstream versions or from local Kubo installations.

And yet... it doesn't. set-target.sh will take a perfectly valid go.mod/sum
files and fuck them up in the end by adding some indirect dependency that
breaks the plugin (on load). I'm sure Go changed behaviour or who knows what.

But for the sake of everyone's time in the future, I'm not even attempting to
fix this magic go mod edit/replace script when I can just manually inspect the
go.mod and set boxo/kubo versions to whatever they need. Anyone is going to have
an easier time manually editing those two lines in the go.mod than figuring out
the magic in the script.